requirements.txt
absl-py==1.2.0
antlr4-python3-runtime==4.9.3
brotlipy @ file:///D:/bld/brotlipy_1648854347500/work
cachetools==5.2.0
certifi @ file:///C:/Windows/TEMP/abs_e9b7158a-aa56-4a5b-87b6-c00d295b01fanefpc8_o/croots/recipe/certifi_1655968940823/work/certifi
cffi @ file:///D:/bld/cffi_1656782940775/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1661170624537/work
colorama==0.4.6
commonmark==0.9.1
connected-components-3d==3.10.2
contourpy==1.0.5
cryptography @ file:///D:/bld/cryptography_1657174142266/work
cycler==0.11.0
DateTime==4.7
-e git+https://github.com/wxdrizzle/deep_kit.git@94fbd8d9f81c0d945f8d7154b821bd782bdd7f61#egg=deep_kit
einops==0.5.0
et-xmlfile==1.1.0
fancycompleter==0.9.1
fonttools==4.37.3
google-auth==2.12.0
google-auth-oauthlib==0.4.6
grpcio==1.49.1
h5py==3.7.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imagecodecs @ file:///D:/bld/imagecodecs_1664236126319/work
imageio==2.22.1
interpolation @ file:///home/conda/feedstock_root/build_artifacts/interpolation_1663795990783/work
joblib==1.2.0
kiwisolver==1.4.4
llvmlite==0.39.1
logger==1.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.0
monai-weekly==1.3.dev2328
networkx @ file:///opt/conda/conda-bld/networkx_1647437648384/work
nibabel==5.0.1
numba @ file:///D:/bld/numba_1664507984908/work
numpy @ file:///D:/bld/numpy_1662889088304/work
oauthlib==3.2.1
omegaconf==2.2.3
opencv-python==4.6.0
openpyxl==3.0.10
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1637239678211/work
pandas==1.5.0
pdbpp==0.10.3
Pillow @ file:///D:/bld/pillow_1660385947107/work
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pydicom==2.3.0
Pygments==2.13.0
pyOpenSSL @ file:///home/conda/feedstock_root/build_artifacts/pyopenssl_1663846997386/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/pyparsing_1652235407899/work
pyreadline==2.1
pyrepl==0.9.0
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil==2.8.2
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1661872987712/work
requests-oauthlib==1.3.1
rich==12.5.1
rsa==4.9
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
seaborn==0.12.1
setuptools-scm==7.1.0
SimpleITK==2.2.0
six==1.16.0
Tempita==0.5.3.dev0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
threadpoolctl==3.1.0
tifffile @ file:///home/conda/feedstock_root/build_artifacts/tifffile_1660549926389/work
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchvision==0.13.1
tqdm==4.65.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1656706066251/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1658789158161/work
Werkzeug==2.2.2
win-inet-pton @ file:///D:/bld/win_inet_pton_1648771897787/work
wmctrl==0.4
zope.interface==5.5.2
